+++ modules/core/Collections/views/collection.php
@@ -46,6 +46,7 @@
                                     <option value="html">Html</option>
                                     <option value="wysiwyg">Html (WYSIWYG)</option>
                                     <option value="code">Code</option>
+                                    <option value="markdown">Markdown</option>
                                     <option value="date">Date</option>
                                     <option value="time">Time</option>
                                     <option value="media">Media</option>